+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Replace `color` and `icon` by `button`. This new component DSL call provides defaults for `button` options when intents render, similar to `label`.
 - Allow blank intents (intents that have no component), useful for creating custom buttons within exposed intents.
 - Accept argument `turbo_frame` in `render_sub_comp`
+- Fix a problem with translations when fastgettext is not installed
 
 ## Steps to take